+++ Documentation/ChangeLog.md
@@ -3,7 +3,15 @@
 All notable changes to this project will be documented here.
 This project adheres to [Semantic Versioning](http://semver.org/)
 
-## UNRELEASED [0.1.0] - 2015-09-08
+## UNRELEASED [0.2.0] - 2015-09-09
+
+### Added
+
+* Toolbox rendering
+* Quick event handling
+
+
+## [0.1.0] - 2015-09-08
 
 ### Notes